format: inference核心模块重构
