+++ app/services/inference/components/__init__.py
@@ -1,0 +1,14 @@
+"""
+推理组件模块
+"""
+
+from .visualizer import DefaultVisualizer
+from .temporal_analyzer import TemporalAnalyzer, DefaultTemporalAnalyzer
+from .component_factory import ComponentFactory
+
+__all__ = [
+    "DefaultVisualizer",
+    "TemporalAnalyzer",
+    "DefaultTemporalAnalyzer",
+    "ComponentFactory",
+]
