+++ app/services/inference/components/temporal_analyzer.py
