+++ app/services/inference/core/__init__.py
@@ -1,0 +1,15 @@
+"""
+推理核心模块
+"""
+
+from .manager import InferenceManager
+from .service import ModelWorkerService
+from .dispatcher import StageAwareDispatcher
+from .factory import create_model_worker_service_from_manager
+
+__all__ = [
+    "InferenceManager",
+    "ModelWorkerService",
+    "StageAwareDispatcher",
+    "create_model_worker_service_from_manager",
+]
