+++ app/services/inference/persistence/__init__.py
@@ -1,0 +1,11 @@
+"""
+推理持久化模块
+
+包含 HLS 段落盘和告警处理逻辑
+"""
+
+# 未来可以将 InferenceManager 中的持久化逻辑提取到此处
+# from .hls_writer import HLSWriter
+# from .alarm_handler import AlarmHandler
+
+__all__ = []
