+++ app/services/inference/workers/__init__.py
@@ -1,0 +1,19 @@
+"""
+Worker 池模块
+"""
+
+from .base import MultiModelWorkerPool
+from .temporal import TemporalWorkerPool, TemporalWorker
+from .visualization import VisualizationWorkerPool, VisualizationWorker, Visualizer
+from .writeback import WriteBackWorkerPool, WriteBackWorker
+
+__all__ = [
+    "MultiModelWorkerPool",
+    "TemporalWorkerPool",
+    "TemporalWorker",
+    "VisualizationWorkerPool",
+    "VisualizationWorker",
+    "Visualizer",
+    "WriteBackWorkerPool",
+    "WriteBackWorker",
+]
