+++ app/services/inference/workers/base.py
