+++ app/services/inference/workers/visualization.py
